better apps structure
