--- xshop/core/api/serializers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-
-from ..models import Product
-
-
-class ProductSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Product
-        fields = ("id", "name", "price", "stock", "added_by")
