+++ xshop/orders/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class OrdersConfig(AppConfig):
+    name = "xshop.orders"
