+++ xshop/orders/models.py
@@ -1,0 +1,28 @@
+from django.db import models
+from django.core.validators import MinValueValidator
+
+from model_utils.models import TimeStampedModel
+
+
+class Order(TimeStampedModel):
+    user = models.ForeignKey(
+        "users.User", on_delete=models.SET_NULL, null=True, blank=True
+    )
+    shop = models.ForeignKey(
+        "shops.Shop", on_delete=models.SET_NULL, null=True, blank=True
+    )
+    paid = models.BooleanField(default=False)
+
+
+class OrderItem(TimeStampedModel):
+    order = models.ForeignKey(Order, on_delete=models.CASCADE)
+    product = models.ForeignKey("products.Product", on_delete=models.CASCADE)
+    quantity = models.SmallIntegerField(validators=[MinValueValidator(1)], default=1)
+
+    @property
+    def unit_price(self):
+        return self.product.price
+
+    @property
+    def total_price(self):
+        return self.product.price * self.quantity
