+++ xshop/products/api/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from ..models import Product
+
+
+class ProductSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Product
+        fields = ("id", "name", "price", "stock", "added_by")
