+++ xshop/products/urls.py
@@ -1,0 +1,3 @@
+app_name = "products"
+
+urlpatterns = []
