+++ xshop/shops/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ShopsConfig(AppConfig):
+    name = "xshop.shops"
